server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
passlib[argon2]>=1.7
pydantic-settings>=2.2
python-multipart>=0.0.9
email-validator>=2.1
redis>=5.0
python-socketio>=5.11
